+++ nestedtensor/csrc/ReduceOps.cpp
@@ -50,68 +50,6 @@ REDUCE_DIM_LIST_FUNC(mean_dim, at::mean, "mean");
 REDUCE_DIM_LIST_FUNC(sum_dim, at::sum, "sum");
 #undef REDUCE_DIM_LIST_FUNC
 
-struct NestedTensorFunction_sum
-    : public torch::autograd::Function<NestedTensorFunction_sum> {
-  static Tensor forward(
-      torch::autograd::AutogradContext* ctx,
-      const Tensor& input_,
-      c10::optional<ScalarType> dtype) {
-    auto input = map_nested_tensor(
-        [](Tensor t) {
-        auto alias = t.alias();
-          alias.requires_grad_();
-          return alias;
-        },
-        input_);
-    auto tensors = flatten(map(
-        [&dtype](at::Tensor tensor) {
-          AutoGradMode autogradmode(true);
-          return at::sum(tensor, dtype);
-        },
-        get_nested_tensor_structure(input)));
-    Tensor result;
-    {
-      AutoGradMode autogradmode(true);
-      if (tensors.size() == 0) {
-        if (dtype) {
-          return at::ones({0}, *dtype);
-        }
-        return at::ones({0});
-      }
-      auto all_tensor = at::stack(tensors);
-      result = at::sum(all_tensor, dtype);
-    }
-    ctx->save_for_backward({result, input});
-    return result.alias();
-  }
-  static torch::autograd::variable_list backward(
-      torch::autograd::AutogradContext* ctx,
-      torch::autograd::variable_list grad_output_) {
-    auto saved = ctx->get_saved_variables();
-    at::Tensor result = saved[0];
-    at::Tensor input = saved[1];
-    at::Tensor grad_output = grad_output_[0];
-    TORCH_CHECK(
-        !grad_output.requires_grad(),
-        "NestedTensor sum doesn't support double backward.");
-    Tensor undef;
-    // TODO:
-    // Flatten constituents and call grad on all of the variable lists at once
-    //
-    at::Tensor tensor = map_nested_tensor(
-        [&](Tensor i) {
-          // return grad_output.expand(i.sizes());
-          return torch::autograd::grad({result}, {i}, {grad_output}, true)[0];
-        },
-        input);
-    return {tensor, undef};
-  }
-};
-
-Tensor NestedTensor_sum(const Tensor& self, c10::optional<ScalarType> dtype) {
-  return NestedTensorFunction_sum::apply(self, dtype);
-}
-
 Tensor NestedTensor_mean(const Tensor& self, c10::optional<ScalarType> dtype) {
   auto tensors = flatten(
       map([&dtype](at::Tensor tensor) { return at::mean(tensor, dtype); },
@@ -142,9 +80,8 @@ Tensor NestedTensor_prod(const Tensor& self, c10::optional<ScalarType> dtype) {
 
 TORCH_LIBRARY_IMPL(aten, PrivateUse1_PreAutograd, m) {
   m.impl_UNBOXED("cumsum", NestedTensor_cumsum);
-  m.impl_UNBOXED("sum.dim_IntList", NestedTensor_sum_dim);
   m.impl_UNBOXED("mean.dim", NestedTensor_mean_dim);
-  m.impl_UNBOXED("sum", NestedTensor_sum);
+  m.impl_UNBOXED("sum.dim_IntList", NestedTensor_sum_dim);
   m.impl_UNBOXED("mean", NestedTensor_mean);
   m.impl_UNBOXED("prod", NestedTensor_prod);
 }
